Best Story Preview; API and Frontend design

+++ HackerNewsDashboard.Common/DTO/ProxyUser.cs
@@ -9,7 +9,7 @@ namespace HackerNewsDashboard.Common.DTO;
 public class ProxyUser
 {
     public int Id { get; set; }
-    public int Created { get; set; }
+    public long Created { get; set; }
     public int Karma { get; set; }
     public string? About { get; set; } = null!;
     public IEnumerable<int>? Submitted { get; set; }
